fix(account) : add validation

+++ wp-content/themes/recruitment-valley/modules/API/features/global/user/UserController.php
@@ -61,9 +61,39 @@ class UserController
 
     public function deleteAccountPermanent(WP_REST_Request $request)
     {
-        require_once(ABSPATH.'wp-admin/includes/user.php');
 
         $userId = $request->user_id;
+        $request = $request->get_params();
+        $userPassword = $request["password"];
+
+        $user = get_user_by( "id", $userId );
+        if(!$user)
+        {
+            return [
+                "status" => 400, 
+                "message" => $this->_message->get("profile.delete.user_not_found"),
+            ];
+        }
+
+        if(UserHelper::is_deleted($user->ID))
+        {
+            return [
+                "status" => 400, 
+                "message" => $this->_message->get("profile.user_deleted"),
+            ];
+        }
+
+        $samePassword = wp_check_password( $userPassword, $user->user_pass, $user->ID );
+        if(!$samePassword)
+        {
+            return [
+                "status" => 400, 
+                "message" => $this->_message->get("profile.delete.password_missmatch"),
+            ];
+        }
+
+        
+        require_once(ABSPATH.'wp-admin/includes/user.php');
         wp_delete_user($userId);
 
         return [
